requirements.txt
# Core dependencies
aiohttp>=3.8.0
pydantic>=2.0.0
click>=8.0.0
pyyaml>=6.0
aiofiles>=22.0.0
colorama>=0.4.6
tabulate>=0.9.0

# Dashboard dependencies (optional)
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.0.0

# Email dependencies (optional)
aiosmtplib>=2.0.0

# Development dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=1.0.0
